+++ packages/components/src/components/date/time/AppTimeInput.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import AppInput from '@/components/input/AppInput.vue'
 import type { Icon } from '@/icons/icons.js'
+import { TimeUtil } from '@/utils/time.util'
 
 const props = withDefaults(defineProps<{
   /**
@@ -55,36 +56,6 @@ const modelValue = defineModel<string | null>({
   required: true,
 })
 
-function formatNumberToTime(time: number): string | null {
-  const timeString = time.toString()
-
-  if (time < 24) {
-    return `${timeString.padStart(2, '0')}:00`
-  }
-  else if (time < 60) {
-    return `00:${timeString.padStart(2, '0')}`
-  }
-  else if (time < 100) {
-    return `01:${(time - 60).toString().padStart(2, '0')}`
-  }
-  else if (timeString.length === 3) {
-    return `${timeString.slice(0, 1)}:${timeString.slice(1)}`
-  }
-  else if (timeString.length === 4) {
-    return `${timeString.slice(0, 2)}:${timeString.slice(2)}`
-  }
-
-  return null
-}
-
-function formatHoursMinutesToTime(hours: number, minutes: number): string | null {
-  if (hours < 24 && minutes < 60) {
-    return `${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}`
-  }
-
-  return null
-}
-
 function onInputBlur(): void {
   if (modelValue.value === null) {
     return
@@ -93,7 +64,7 @@ function onInputBlur(): void {
   const time = modelValue.value
 
   if (/^\d+$/.test(time) && time.length <= 4) {
-    modelValue.value = formatNumberToTime(Number.parseInt(time))
+    modelValue.value = TimeUtil.formatNumberToTime(Number.parseInt(time))
 
     return
   }
@@ -101,7 +72,7 @@ function onInputBlur(): void {
   if (/^\d{1,2}h$/.test(time)) {
     const hours: number = Number.parseInt(time.replace('h', ''))
 
-    modelValue.value = formatHoursMinutesToTime(hours, 0)
+    modelValue.value = TimeUtil.formatHoursMinutesToTime(hours, 0)
 
     return
   }
@@ -109,7 +80,7 @@ function onInputBlur(): void {
   if (/^\d{1,2}m$/.test(time)) {
     const minutes: number = Number.parseInt(time.replace('m', ''))
 
-    modelValue.value = formatHoursMinutesToTime(0, minutes)
+    modelValue.value = TimeUtil.formatHoursMinutesToTime(0, minutes)
 
     return
   }
@@ -119,7 +90,7 @@ function onInputBlur(): void {
     const hours: number = Number.parseInt(timeParts[0] ?? '0')
     const minutes: number = Number.parseInt(timeParts[1] ?? '0')
 
-    modelValue.value = formatHoursMinutesToTime(hours, minutes)
+    modelValue.value = TimeUtil.formatHoursMinutesToTime(hours, minutes)
 
     return
   }
@@ -129,7 +100,7 @@ function onInputBlur(): void {
     const hours: number = Number.parseInt(timeParts[0] ?? '0')
     const minutes: number = Number.parseInt(timeParts[1] ?? '0')
 
-    modelValue.value = formatHoursMinutesToTime(hours, minutes)
+    modelValue.value = TimeUtil.formatHoursMinutesToTime(hours, minutes)
 
     return
   }
